Discord works, even without firebase or timers
This includes DB functions like birthday and time. These will run locally in "development" mode until the Firebase connection is enabled
