+++ src/discord/commands/guild/m_report.js
@@ -6,12 +6,9 @@ const { ApplicationCommandType } = require('discord-api-types/v9');
 const { ContextMenuCommandBuilder } = require('@discordjs/builders');
 const { stripIndents } = require('common-tags/lib');
 const logger = require('../../../global/utils/logger');
-const template = require('../../utils/embed-template');
 const mod = require('./mod');
 
 const PREFIX = path.parse(__filename).name;
-
-const embed = template.embedTemplate();
 
 let actor = {};
 let target = {};
