+++ src/discord/commands/guild/remindme.js
@@ -45,48 +45,14 @@ module.exports = {
     }
 
     // Load actor data
-    const set = await setUserInfo(actorFbid, actorData);
-
-    logger.debug(`[${PREFIX}] global.userDb: ${JSON.stringify(global.userDb)}`);
-    const userDb = [];
-    if (Object.keys(global.userDb).length > 0) {
-      global.userDb.forEach(doc => {
-        if (doc.key === actorFbid) {
-          userDb.push({
-            key: doc.key,
-            value: actorData,
-          });
-          logger.debug(`[${PREFIX}] Updated actor in userDb`);
-        } else {
-          userDb.push({
-            key: doc.key,
-            value: doc.value,
-          });
-        }
-      });
-    } else {
-      const keyString = Math.random().toString(36).substring(2, 10);
-      userDb.push({
-        // Get random string of 8 characters
-        key: keyString,
-        value: actorData,
-      });
-    }
-    Object.assign(global, { userDb });
-    logger.debug(`[${PREFIX}] Updated global user data.`);
+    await setUserInfo(actorFbid, actorData);
 
     logger.debug(`[${PREFIX}] userDb: ${JSON.stringify(global.userDb)}`);
 
     const timeBetween = reminderDatetime - new Date();
 
     const embed = template.embedTemplate()
-      .setDescription(`In ${ms(timeBetween, { long: true })} I will remind you: ${reminder}`)
-
-    if (!set) {
-      embed.addFields(
-        { name: 'Warning', value: 'Could not connec to firebase', inline: true },
-      );
-    }
+      .setDescription(`In ${ms(timeBetween, { long: true })} I will remind you: ${reminder}`);
 
     interaction.reply({ embeds: [embed], ephemeral: true });
