+++ src/discord/events/guild-member-update.js
@@ -138,24 +138,6 @@ module.exports = {
 
           // Load actor data
           await setUserInfo(actorFbid, actorData);
-
-          const userDb = [];
-          global.userDb.forEach(doc => {
-            if (doc.key === actorFbid) {
-              userDb.push({
-                key: doc.key,
-                value: actorData,
-              });
-              logger.debug(`[${PREFIX}] Updated actor in userDb`);
-            } else {
-              userDb.push({
-                key: doc.key,
-                value: doc.value,
-              });
-            }
-          });
-          Object.assign(global, { userDb });
-          logger.debug(`[${PREFIX}] Updated global user data.`);
         }
       }
     }
