+++ src/discord/utils/community-mod.js
@@ -162,24 +162,6 @@ module.exports = {
       // Load actor data
       await setUserInfo(actorFbid, actorData);
 
-      const userDb = [];
-      global.userDb.forEach(doc => {
-        if (doc.key === actorFbid) {
-          userDb.push({
-            key: doc.key,
-            value: actorData,
-          });
-          logger.debug(`[${PREFIX}] Updated actor in userDb`);
-        } else {
-          userDb.push({
-            key: doc.key,
-            value: doc.value,
-          });
-        }
-      });
-      Object.assign(global, { userDb });
-      logger.debug(`[${PREFIX}] Updated global user data.`);
-
       return reaction.message.reply(stripIndents`
         Hey ${moderatorRole}! ${target.user.username} was community banned for this, please review!
